move beans from util to a not excluded package
IntelliJ powerfull refactoring has done it.
