+++ core/src/test/java/com/opensymphony/xwork2/AnnotatedTestBean.java
@@ -19,8 +19,7 @@
 package com.opensymphony.xwork2;
 
 import com.opensymphony.xwork2.conversion.annotations.TypeConversion;
-import com.opensymphony.xwork2.conversion.impl.FooBarConverter;
-import com.opensymphony.xwork2.util.Bar;
+import com.opensymphony.xwork2.beans.util.Bar;
 import com.opensymphony.xwork2.validator.annotations.IntRangeFieldValidator;
 import com.opensymphony.xwork2.validator.annotations.RequiredStringValidator;
 import com.opensymphony.xwork2.validator.annotations.Validations;
