+++ core/src/test/java/com/opensymphony/xwork2/beans/util/BarJunior.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.opensymphony.xwork2.util;
+package com.opensymphony.xwork2.beans.util;
 
 public class BarJunior extends Bar {
 }
