+++ core/src/test/java/com/opensymphony/xwork2/beans/util/Dog.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.opensymphony.xwork2.util;
+package com.opensymphony.xwork2.beans.util;
 
 import java.io.Serializable;
