+++ core/src/test/java/com/opensymphony/xwork2/conversion/impl/XWorkConverterTest.java
@@ -22,10 +22,10 @@ import com.opensymphony.xwork2.*;
 import com.opensymphony.xwork2.ognl.OgnlValueStack;
 import com.opensymphony.xwork2.test.ModelDrivenAction2;
 import com.opensymphony.xwork2.test.User;
-import com.opensymphony.xwork2.util.Bar;
-import com.opensymphony.xwork2.util.Cat;
-import com.opensymphony.xwork2.util.Foo;
-import com.opensymphony.xwork2.util.FurColor;
+import com.opensymphony.xwork2.beans.util.Bar;
+import com.opensymphony.xwork2.beans.util.Cat;
+import com.opensymphony.xwork2.beans.util.Foo;
+import com.opensymphony.xwork2.beans.util.FurColor;
 import com.opensymphony.xwork2.util.reflection.ReflectionContextState;
 import ognl.OgnlException;
 import ognl.OgnlRuntime;
