+++ core/src/test/java/com/opensymphony/xwork2/test/DataAware.java
@@ -18,7 +18,7 @@
  */
 package com.opensymphony.xwork2.test;
 
-import com.opensymphony.xwork2.util.Bar;
+import com.opensymphony.xwork2.beans.util.Bar;
 
 
 /**
