+++ core/src/test/java/com/opensymphony/xwork2/test/SimpleAction3.java
@@ -19,7 +19,7 @@
 package com.opensymphony.xwork2.test;
 
 import com.opensymphony.xwork2.SimpleAction;
-import com.opensymphony.xwork2.util.Bar;
+import com.opensymphony.xwork2.beans.util.Bar;
 
 
 /**
