+++ core/src/test/resources/com/opensymphony/xwork2/conversion/impl/test-xwork-conversion.properties
@@ -16,4 +16,4 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-com.opensymphony.xwork2.util.Bar=com.opensymphony.xwork2.conversion.impl.FooBarConverter
+com.opensymphony.xwork2.beans.util.Bar=com.opensymphony.xwork2.conversion.impl.FooBarConverter
