+++ core/src/test/resources/com/opensymphony/xwork2/util/Cat-conversion.properties
@@ -16,5 +16,5 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-Collection_kittens = com.opensymphony.xwork2.util.Cat
+Collection_kittens = com.opensymphony.xwork2.beans.util.Cat
 foo.number = com.opensymphony.xwork2.conversion.impl.FooNumberConverter
