+++ core/src/test/resources/com/opensymphony/xwork2/util/Foo-conversion.properties
@@ -17,13 +17,13 @@
 # under the License.
 #
 bar=com.opensymphony.xwork2.conversion.impl.FooBarConverter
-Element_cats=com.opensymphony.xwork2.util.Cat
-Element_moreCats=com.opensymphony.xwork2.util.Cat
-Element_catMap=com.opensymphony.xwork2.util.Cat
+Element_cats=com.opensymphony.xwork2.beans.util.Cat
+Element_moreCats=com.opensymphony.xwork2.beans.util.Cat
+Element_catMap=com.opensymphony.xwork2.beans.util.Cat
 Key_anotherCatMap=java.lang.Long
-Element_anotherCatMap=com.opensymphony.xwork2.util.Cat
+Element_anotherCatMap=com.opensymphony.xwork2.beans.util.Cat
 KeyProperty_barCollection=id
-Element_barCollection=com.opensymphony.xwork2.util.Bar
+Element_barCollection=com.opensymphony.xwork2.beans.util.Bar
 KeyProperty_barList=id
-Element_barList=com.opensymphony.xwork2.util.Bar
+Element_barList=com.opensymphony.xwork2.beans.util.Bar
