+++ core/src/test/resources/com/opensymphony/xwork2/util/MyBeanAction-conversion.properties
@@ -17,10 +17,10 @@
 # under the License.
 #
 KeyProperty_beanList=id
-Element_beanList=com.opensymphony.xwork2.util.MyBean
+Element_beanList=com.opensymphony.xwork2.beans.util.MyBean
 CreateIfNull_beanList=true
 
 Key_beanMap=java.lang.Long
 KeyProperty_beanMap=id
-Element_beanMap=com.opensymphony.xwork2.util.MyBean
+Element_beanMap=com.opensymphony.xwork2.beans.util.MyBean
