+++ core/src/test/resources/com/opensymphony/xwork2/util/Tiger-conversion.properties
@@ -16,4 +16,4 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-Collection_dogs = com.opensymphony.xwork2.util.Dog
+Collection_dogs = com.opensymphony.xwork2.beans.util.Dog
